+++ example/lib/typedef_example.dart
@@ -1,0 +1,17 @@
+import 'package:mek_data_class/mek_data_class.dart';
+
+part 'typedef_example.g.dart';
+
+typedef JsonMap<T> = Map<String, T>;
+typedef NullableJsonMap<T> = Map<String?, T>;
+
+@DataClass()
+class Extra with _$Extra {
+  final JsonMap<int>? jsonMap;
+  final NullableJsonMap<double?> nullableJsonMap;
+
+  const Extra({
+    required this.jsonMap,
+    required this.nullableJsonMap,
+  });
+}