+++ mek_data_class_generator/CHANGELOG.md
@@ -1,3 +1,5 @@
+- feat: Added support for type aliases (typedef)
+
 ## 0.2.0
 - Added `DataClass.createFieldsClass` to create a class that contains the names of the fields of the Data Class